+++ idTech4Amm/src/main/java/com/karin/idTech4Amm/sys/Constants.java
@@ -13,7 +13,7 @@ import java.util.Date;
 public final class Constants
 {
     public static final int CONST_UPDATE_RELEASE = 39;
-    public static final String CONST_RELEASE = "2024-02-18";
+    public static final String CONST_RELEASE = "2024-04-10";
     public static final String CONST_EMAIL = "beyondk2000@gmail.com";
     public static final String CONST_DEV = "Karin";
     public static final String CONST_CODE = "Harmattan";
@@ -28,11 +28,13 @@ public final class Constants
 	public static final String CONST_CHECK_FOR_UPDATE_URL = "https://raw.githubusercontent.com/glKarin/com.n0n3m4.diii4a/master/CHECK_FOR_UPDATE.json";
     public static final String CONST_LICENSE_URL = "https://raw.githubusercontent.com/glKarin/com.n0n3m4.diii4a/master/LICENSE";
 	public static final String[] CONST_CHANGES = {
-            "Add `LibreCoop` mod of DOOM3 support, game data directory named `librecoop`. More view in`" + TextHelper.GenLinkText("https://www.moddb.com/mods/librecoop-dhewm3-coop", "LibreCoop") + "`.",
-            "Add `Quake II` support, game data directory named `baseq2`. More view in`" + TextHelper.GenLinkText("https://store.steampowered.com/app/2320/Quake_II/", "Quake II") + "`.",
-            "Add `Quake III Arena` support, game data directory named `baseq3`. More view in`" + TextHelper.GenLinkText("https://store.steampowered.com/app/2200/Quake_III_Arena/", "Quake III Arena") + "`.",
-            "Add `Return to Castle Wolfenstein` support, game data directory named `main`. More view in`" + TextHelper.GenLinkText("https://www.moddb.com/games/return-to-castle-wolfenstein", "Return to Castle Wolfenstein") + "`.",
-            "Add `The dark mod` support, game data directory named `darkmod`. More view in`" + TextHelper.GenLinkText("https://www.thedarkmod.com", "The Dark Mod") + "`.",
+            "Support perforated surface shadow in shadow mapping.",
+            "Add `LibreCoop` mod of DOOM3 support, game data directory named `librecoop`. More view in `" + TextHelper.GenLinkText("https://www.moddb.com/mods/librecoop-dhewm3-coop", "LibreCoop") + "`.",
+            "Add `Quake II` support, game data directory named `baseq2`. More view in `" + TextHelper.GenLinkText("https://store.steampowered.com/app/2320/Quake_II/", "Quake II") + "`.",
+            "Add `Quake III Arena` support, game data directory named `baseq3`; Add `Quake III Team Arena` support, game data directory named `missionpack`. More view in `" + TextHelper.GenLinkText("https://store.steampowered.com/app/2200/Quake_III_Arena/", "Quake III Arena") + "`.",
+            "Add `Return to Castle Wolfenstein` support, game data directory named `main`. More view in `" + TextHelper.GenLinkText("https://www.moddb.com/games/return-to-castle-wolfenstein", "Return to Castle Wolfenstein") + "`.",
+            "Add `The dark mod` support, game data directory named `darkmod`. More view in `" + TextHelper.GenLinkText("https://www.thedarkmod.com", "The Dark Mod") + "`.",
+            "Add a on-screen button theme.",
 	};
 
     public static long GetBuildTimestamp()